+++ orthoseg/load_sampleprojects.py
@@ -2,12 +2,14 @@
 
 import argparse
 import logging
+import shutil
 import sys
+import tempfile
+import urllib.request
+import zipfile
 from pathlib import Path
 
 import gdown
-
-from orthoseg.util import git_downloader
 
 # Get a logger...
 logger = logging.getLogger(__name__)
@@ -59,43 +61,61 @@ def load_sampleprojects(dest_dir: Path, ssl_verify: bool | str = True):
             this will be used. In corporate networks using a proxy server this is often
             needed to avoid CERTIFICATE_VERIFY_FAILED errors. Defaults to True.
     """
+    dest_dir = dest_dir.expanduser()
     dest_dir_full = dest_dir / "sample_projects"
     if dest_dir_full.exists():
         raise ValueError(f"Destination directory already exists: {dest_dir_full}")
 
     # Download
     print(f"Start download of sample projects to {dest_dir_full!s}")
-    git_downloader.download(
-        repo_url="https://github.com/orthoseg/orthoseg/tree/main/sample_projects",
-        output_dir=dest_dir,
-        ssl_verify=ssl_verify,
-    )
+    with tempfile.TemporaryDirectory(ignore_cleanup_errors=True) as tmp_dir:
+        # Download the sample projects to a temporary directory first
+        url = "https://github.com/orthoseg/orthoseg_sample_projects/archive/refs/tags/v0.1.0.zip"
+        tmp_zip_path = Path(tmp_dir) / "sample_projects.zip"
+        urllib.request.urlretrieve(url, tmp_zip_path)
+
+        # Unzip the downloaded file
+        tmp_proj_dir = Path(tmp_dir) / "sample_projects_tmp"
+        with zipfile.ZipFile(tmp_zip_path, "r") as zip_file:
+            zip_file.extractall(tmp_proj_dir)
+        # Get the single subdirectory in the unzipped folder
+        subdirs = [d for d in tmp_proj_dir.iterdir() if d.is_dir()]
+        if len(subdirs) != 1:
+            raise ValueError(
+                f"Expected exactly one subdirectory, but found {len(subdirs)}"
+            )
+
+        # Move the unzipped sample projects to the destination directory
+        dest_dir.mkdir(parents=True, exist_ok=True)
+        shutil.move(subdirs[0] / "sample_projects", dest_dir)
+
     print("Download finished")
+
     print("Start download of footballfields pretrained neural net")
-    footballfields_model_dir = dest_dir_full / "footballfields/models"
-    footballfields_model_dir.mkdir(parents=True, exist_ok=True)
-    model_hdf5_path = footballfields_model_dir / "footballfields_01_0.92512_242.hdf5"
-    if model_hdf5_path.exists() is False:
+    verify = True if ssl_verify is None else ssl_verify
+    model_dir = dest_dir_full / "footballfields/models"
+    model_dir.mkdir(parents=True, exist_ok=True)
+
+    model_hdf5_path = model_dir / "footballfields_01_0.97392_201.hdf5"
+    if not model_hdf5_path.exists():
         gdown.download(
-            id="1XmAenCW6K_RVwqC6xbkapJ5ws-f7-QgH",
+            id="1UlNorZ74ADCr3pL4MCJ_tnKRNoeZX79g",
             output=str(model_hdf5_path),
-            verify=ssl_verify,
+            verify=verify,
         )
-    model_hyperparams_path = (
-        footballfields_model_dir / "footballfields_01_hyperparams.json"
-    )
-    if model_hyperparams_path.exists() is False:
+    model_hyperparams_path = model_dir / "footballfields_01_hyperparams.json"
+    if not model_hyperparams_path.exists():
         gdown.download(
-            id="1umxcd4RkB81sem9PdIpLoWeiIW8ga1u7",
+            id="1NwrVVjx9IsjvaioQ4-bkPMrq7S6HeWIo",
             output=str(model_hyperparams_path),
-            verify=ssl_verify,
+            verify=verify,
         )
-    model_modeljson_path = footballfields_model_dir / "footballfields_01_model.json"
-    if model_modeljson_path.exists() is False:
+    model_modeljson_path = model_dir / "footballfields_01_model.json"
+    if not model_modeljson_path.exists():
         gdown.download(
-            id="16qe8thBTrO3dFfLMU1T22gWcfHVXt8zQ",
+            id="1LNPLypM5in3aZngBKK_U4Si47Oe97ZWN",
             output=str(model_modeljson_path),
-            verify=ssl_verify,
+            verify=verify,
         )
     print("Download finished")