+++ tests/test_helper.py
@@ -17,8 +17,8 @@ sampleprojects_dir = Path(__file__).resolve().parent.parent / "sample_projects"
 
 class SampleProjectFootball:
     project_dir = sampleprojects_dir / "footballfields"
-    predict_config_path = project_dir / "footballfields_BEFL-2019_test.ini"
-    train_config_path = project_dir / "footballfields_train_test.ini"
+    predict_config_path = project_dir / "footballfields_BEFL-2019.ini"
+    config_path = project_dir / "footballfields.ini"
 
     @staticmethod
     def download_model(dst_dir):