+++ tests/test_postprocess.py
@@ -30,7 +30,7 @@ def test_postprocess_error_handling():
     """Force an error so the general error handler in postprocess is tested."""
     with pytest.raises(
         RuntimeError,
-        match="ERROR in postprocess for footballfields_BEFL-2019_test on UNEXISTING",
+        match="ERROR in postprocess for footballfields_BEFL-2019 on UNEXISTING",
     ):
         postprocess(
             config_path=test_helper.SampleProjectFootball.predict_config_path,